requirements.txt
pyqt5
pycups
PyPDF2
